% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{citationTabUI}
\alias{citationTabUI}
\title{Citation Tab UI}
\usage{
citationTabUI(citations)
}
\arguments{
\item{citations}{A list of citations generated by cite_r_package or cite_bibentry.}
}
\description{
Generates the user interface for tabbed citations in shiny.
}
