% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{citationUI}
\alias{citationUI}
\title{Citation UI}
\usage{
citationUI(citations, title = "References")
}
\arguments{
\item{citations}{A list of citations generated by cite_r_package or cite_bibentry.}

\item{title}{Title for the user interface element}
}
\description{
Generates the user interface for citations in shiny.
}
