% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_bibentry.R
\name{cite_bibentry}
\alias{cite_bibentry}
\title{Generate a citation from a bibentry}
\usage{
cite_bibentry(bibentry, use = NULL)
}
\arguments{
\item{bibentry}{A bibentry object}

\item{use}{Optional usage to add to the citation (e.g. 'Visualisation', 'Data Processing')}
}
\description{
Generate a citation for use with schite using a bibentry object.
}
