% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_r_package.R
\name{cite_r_package}
\alias{cite_r_package}
\title{Cite an R package}
\usage{
cite_r_package(pkg, use = NULL)
}
\arguments{
\item{pkg}{The name of the package}

\item{use}{Optional usage to add to the citation (e.g. 'Visualisation', 'Data Processing')}
}
\description{
Generate a citation for use with schite using the CITATION file
within an R package.
}
